Remove dead code from the base-model migration
Deletes the now-unused LlamaPromptRenderer (the OSS path renders via BaseCompletionPromptRenderer after the cut). Moves the still-live LlamaPromptCacheHintTracker tests into their own file and drops the dead instruct-renderer tests from ExtendedContext/CustomRules/LanguageSupport.

Also repoints the two prompt-content assertions (clipboard label, reference-notes label) to the base-prompt strings, fixing the two tests that were failing #497's CI, and updates stale comments that referenced the removed renderer.
