+++ CotabbyTests/LlamaPromptCacheHintTrackerTests.swift
@@ -1,0 +1,108 @@
+import CoreGraphics
+import XCTest
+@testable import Cotabby
+
+/// Tests for `LlamaPromptCacheHintTracker`, the conservative byte-prefix hint the llama engine
+/// passes into the runtime to reuse KV state across keystrokes. Pure-function and deterministic:
+/// the tracker only advertises reuse for the same focused field and sampling fingerprint.
+final class LlamaPromptCacheHintTrackerTests: XCTestCase {
+
+    // MARK: - cache hints
+
+    func test_cacheHint_nilBeforeSuccessfulRequestIsRecorded() {
+        var tracker = LlamaPromptCacheHintTracker()
+
+        XCTAssertNil(tracker.cachedPrefixBytes(for: makeRequest(prompt: "hello")))
+    }
+
+    func test_cacheHint_returnsCommonPrefixBytesForSameFocusedField() {
+        var tracker = LlamaPromptCacheHintTracker()
+        tracker.recordSuccessfulRequest(makeRequest(prompt: "hello"))
+
+        XCTAssertEqual(
+            tracker.cachedPrefixBytes(for: makeRequest(prompt: "hello!")),
+            "hello".utf8.count
+        )
+    }
+
+    func test_cacheHint_invalidatesWhenFocusedFieldChanges() {
+        var tracker = LlamaPromptCacheHintTracker()
+        tracker.recordSuccessfulRequest(makeRequest(prompt: "hello", elementIdentifier: "field-a"))
+
+        XCTAssertNil(
+            tracker.cachedPrefixBytes(for: makeRequest(prompt: "hello!", elementIdentifier: "field-b"))
+        )
+    }
+
+    func test_cacheHint_prefersStableInputFrameOverUnstableElementIdentifier() {
+        var tracker = LlamaPromptCacheHintTracker()
+        let fieldFrame = CGRect(x: 10, y: 20, width: 300, height: 44)
+        tracker.recordSuccessfulRequest(
+            makeRequest(prompt: "hello", elementIdentifier: "field-a", inputFrameRect: fieldFrame)
+        )
+
+        XCTAssertEqual(
+            tracker.cachedPrefixBytes(
+                for: makeRequest(prompt: "hello!", elementIdentifier: "field-b", inputFrameRect: fieldFrame)
+            ),
+            "hello".utf8.count
+        )
+    }
+
+    func test_cacheHint_invalidatesWhenSamplingFingerprintChanges() {
+        var tracker = LlamaPromptCacheHintTracker()
+        tracker.recordSuccessfulRequest(makeRequest(prompt: "hello", topK: 20))
+
+        XCTAssertNil(tracker.cachedPrefixBytes(for: makeRequest(prompt: "hello!", topK: 40)))
+    }
+
+    // MARK: - helpers
+
+    private func makeRequest(
+        prompt: String,
+        elementIdentifier: String = "field",
+        topK: Int = 20,
+        inputFrameRect: CGRect? = nil
+    ) -> SuggestionRequest {
+        let snapshot = FocusedInputSnapshot(
+            applicationName: "TestApp",
+            bundleIdentifier: "com.example.TestApp",
+            processIdentifier: 123,
+            elementIdentifier: elementIdentifier,
+            role: "AXTextField",
+            subrole: nil,
+            caretRect: .zero,
+            inputFrameRect: inputFrameRect,
+            caretSource: "test",
+            caretQuality: .exact,
+            observedCharWidth: nil,
+            precedingText: prompt,
+            trailingText: "",
+            selection: NSRange(location: prompt.count, length: 0),
+            isSecure: false
+        )
+        let context = FocusedInputContext(snapshot: snapshot, generation: 1)
+
+        return SuggestionRequest(
+            context: context,
+            prefixText: prompt,
+            prompt: prompt,
+            generation: context.generation,
+            maxPredictionTokens: 8,
+            temperature: 0.1,
+            topK: topK,
+            topP: 0.7,
+            minP: 0.08,
+            repetitionPenalty: 1.05,
+            randomSeed: 42,
+            maxSuffixCharacters: 192,
+            completionLengthInstruction: "Return only the next few words.",
+            userName: nil,
+            customRules: [],
+            languageInstruction: nil,
+            clipboardContext: nil,
+            visualContextSummary: nil,
+            isMultiLineEnabled: false
+        )
+    }
+}
